+++ test/OpenTelemetry.Tests/Shared/TestExportProcessor.cs
@@ -1,0 +1,36 @@
+// <copyright file="TestExportProcessor.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using System.Collections.Generic;
+
+namespace OpenTelemetry.Tests
+{
+    internal class TestExportProcessor<T> : BaseExportProcessor<T>
+        where T : class
+    {
+        public List<T> ExportedItems = new List<T>();
+
+        public TestExportProcessor(BaseExporter<T> exporter)
+            : base(exporter)
+        {
+        }
+
+        public override void OnExport(T data)
+        {
+            this.ExportedItems.Add(data);
+        }
+    }
+}